requirements.txt
aiohttp==3.8.0
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
datasets==1.15.1
decorator==5.1.0
dill==0.3.4
filelock==3.3.2
frozenlist==1.2.0
fsspec==2021.11.0
huggingface-hub==0.1.2
idna==3.3
imbalanced-learn==0.8.1
ipython==7.29.0
iteration-utilities==0.11.0
jedi==0.18.0
joblib==1.1.0
matplotlib-inline==0.1.3
multidict==5.2.0
multiprocess==0.70.12.2
numpy==1.21.4
packaging==21.2
pandas==1.3.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pyarrow==6.0.0
pycodestyle==2.8.0
Pygments==2.10.0
pyjnius==1.4.1
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==2.8.2
python-rdrsegmenter==0.1.1
pytz==2021.3
pyvi==0.1.1
PyYAML==6.0
rank-bm25==0.2.1
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn-crfsuite==0.3.6
tabulate==0.8.9
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.10.0
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
xxhash==2.0.2
yarl==1.7.2
